+++ ccramic/callbacks/pixel_level_callbacks.py
@@ -9,7 +9,8 @@ from ccramic.inputs.pixel_level_inputs import (
     wrap_canvas_in_loading_screen_for_large_images,
     invert_annotations_figure,
     set_range_slider_tick_markers,
-    generate_canvas_legend_text)
+    generate_canvas_legend_text,
+    set_x_axis_placement_of_scalebar)
 from ccramic.parsers.pixel_level_parsers import (
     # populate_upload_dict,
     FileParser,
@@ -955,7 +956,7 @@ def init_pixel_level_callbacks(dash_app, tmpdirname, authentic_id, app_config):
                        State('toggle-canvas-scalebar', 'value'),
                        Input('mask-blending-slider', 'value'),
                        Input('add-mask-boundary', 'value'),
-                       Input('channel-order', 'data'),
+                       State('channel-order', 'data'),
                        State('legend-size-slider', 'value'),
                        Input('add-cell-id-mask-hover', 'value'),
                        State('pixel-size-ratio', 'value'),
@@ -971,6 +972,7 @@ def init_pixel_level_callbacks(dash_app, tmpdirname, authentic_id, app_config):
                        State('imported-cluster-frame', 'data'),
                        Input('cluster-annotation-type', 'value'),
                        Input('btn-download-canvas-tiff', 'n_clicks'),
+                       State('custom-scale-val', 'value'),
                        prevent_initial_call=True)
     # @cache.memoize())
     def render_canvas_from_layer_mask_hover_change(canvas_layers, currently_selected,
@@ -983,7 +985,7 @@ def init_pixel_level_callbacks(dash_app, tmpdirname, authentic_id, app_config):
                                                 pixel_ratio, invert_annot, overlay_grid, legend_orientation,
                                                 global_apply_filter, global_filter_type, global_filter_val,
                                                 global_filter_sigma, apply_cluster_on_mask, cluster_assignments_dict,
-                                                cluster_frame, cluster_type, download_canvas_tiff):
+                                                cluster_frame, cluster_type, download_canvas_tiff, custom_scale_val):
 
         """
         Update the canvas from a layer dictionary update (The cache dictionary containing the modified image layers
@@ -999,7 +1001,7 @@ def init_pixel_level_callbacks(dash_app, tmpdirname, authentic_id, app_config):
                                                   "global-sigma-val-filter"] and not global_apply_filter
         channel_order_same = ctx.triggered_id in ["channel-order"] and channel_order == currently_selected
         if canvas_layers is not None and currently_selected is not None and blend_colour_dict is not None and \
-                data_selection is not None and len(currently_selected) > 0 and len(canvas_children) > 0 and \
+                data_selection is not None and currently_selected and len(canvas_children) > 0 and \
                 len(channel_order) > 0 and not global_not_enabled and not channel_order_same:
             cur_graph = strip_invalid_shapes_from_graph_layout(cur_graph)
             # try:
@@ -1016,7 +1018,7 @@ def init_pixel_level_callbacks(dash_app, tmpdirname, authentic_id, app_config):
                  legend_text, toggle_scalebar, legend_size, toggle_legend, add_cell_id_hover,
                  show_each_channel_intensity, raw_data_dict, aliases, global_apply_filter,
                 global_filter_type, global_filter_val, global_filter_sigma,
-                apply_cluster_on_mask, cluster_assignments_dict, cluster_frame, cluster_type)
+                apply_cluster_on_mask, cluster_assignments_dict, cluster_frame, cluster_type, custom_scale_val)
                 fig = canvas.generate_canvas()
                 if cluster_type == 'mask' or not apply_cluster_on_mask:
                     fig = CanvasLayout(fig)
@@ -1033,20 +1035,9 @@ def init_pixel_level_callbacks(dash_app, tmpdirname, authentic_id, app_config):
                     canvas_tiff = dcc.send_file(output_current_canvas_as_tiff(canvas_image=canvas.get_image(),
                                                 dest_dir=dest_path))
                 return fig, canvas_tiff
-                # else
-                #     fig = CanvasLayout(cur_graph)
-                #     if apply_cluster_on_mask:
-                #         fig = fig.add_cluster_annotations_as_circles(mask_config[mask_selection]["raw"],
-                #                 pd.DataFrame(cluster_frame), cluster_assignments_dict, data_selection)
-                #     elif not apply_cluster_on_mask:
-                #         fig = fig.remove_cluster_annotation_shapes()
-                #     else:
-                #         fig = fig.get_fig()
-                #     return go.Figure(fig), dash.no_update
             except (ValueError, AttributeError, KeyError, IndexError):
                 raise PreventUpdate
         #TODO: this step can be used to keep the current ui revision if a new ROI is selected with the same dimensions
-
         # elif currently_selected is not None and 'shapes' not in cur_graph_layout:
         #     fig = cur_graph if cur_graph is not None else go.Figure()
         #     if 'data' in fig:
@@ -1080,11 +1071,18 @@ def init_pixel_level_callbacks(dash_app, tmpdirname, authentic_id, app_config):
                        State('set-y-auto-bound', 'value'),
                        State('window_config', 'data'),
                        Input('activate-coord', 'n_clicks'),
-                       State('pixel-size-ratio', 'value'),
+                       State('data-collection', 'value'),
+                       State('uploaded_dict', 'data'),
+                       Input('custom-scale-val', 'value'),
+                       Input('pixel-size-ratio', 'value'),
+                       State('toggle-canvas-scalebar', 'value'),
+                       State('legend-size-slider', 'value'),
+                       State('invert-annotations', 'value'),
                        prevent_initial_call=True)
     # @cache.memoize())
-    def render_canvas_from_coord_change(cur_graph, cur_graph_layout, x_request, y_request, current_window,
-                             nclicks_coord, pixel_ratio):
+    def render_canvas_from_coord_or_zoom_change(cur_graph, cur_graph_layout, x_request, y_request, current_window,
+                             nclicks_coord, data_selection, image_dict, custom_scale_val, pixel_ratio,
+                            toggle_scalebar, legend_size, invert_annot):
 
         """
         Update the annotation canvas when the zoom or custom coordinates are requested.
@@ -1092,79 +1090,30 @@ def init_pixel_level_callbacks(dash_app, tmpdirname, authentic_id, app_config):
         # bad_update = cur_graph_layout in [{"autosize": True}]
 
         # update the scale bar with and without zooming
-        if cur_graph is not None:
+        if None not in (cur_graph, cur_graph_layout, data_selection):
             cur_graph = strip_invalid_shapes_from_graph_layout(cur_graph)
             pixel_ratio = pixel_ratio if pixel_ratio is not None else 1
-            if ctx.triggered_id == "annotation_canvas":
+            if ctx.triggered_id not in ["activate-coord"]:
                 try:
-                    # fig = go.Figure(cur_graph)
-                    # # find the text annotation that has um in the text and the correct location
-                    # for annotations in cur_graph['layout']['annotations']:
-                    #     # if 'μm' in annotations['text'] and annotations['y'] == 0.06:
-                    #     if annotations['y'] == 0.06:
-                    #         if cur_graph_layout not in [{'autosize': True}]:
-                    #             x_range_high = 0
-                    #             x_range_low = 0
-                    #             # use different variables depending on how the ranges are written in the dict
-                    #             # IMP: the variables will be written differently after a tab change
-                    #             if 'xaxis' in cur_graph['layout']:
-                    #                 high = max(cur_graph['layout']['xaxis']['range'][1],
-                    #                        cur_graph['layout']['xaxis']['range'][0])
-                    #                 low = min(cur_graph['layout']['xaxis']['range'][1],
-                    #                       cur_graph['layout']['xaxis']['range'][0])
-                    #                 x_range_high = math.ceil(int(high))
-                    #                 x_range_low = math.floor(int(low))
-                    #             elif 'xaxis.range[0]' and 'xaxis.range[1]' in cur_graph_layout:
-                    #                 high = max(cur_graph_layout['xaxis.range[1]'],
-                    #                        cur_graph_layout['xaxis.range[0]'])
-                    #                 low = min(cur_graph_layout['xaxis.range[1]'],
-                    #                       cur_graph_layout['xaxis.range[0]'])
-                    #                 x_range_high = math.ceil(int(high))
-                    #                 x_range_low = math.ceil(int(low))
-                    #
-                    #             assert x_range_high >= x_range_low
-                    #             # assert that all values must be above 0 for the scale value to render during panning
-                    #             # assert all([elem >=0 for elem in cur_graph_layout.values() if isinstance(elem, float)])
-                    #             scale_val = int(float(math.ceil(int(0.075 * (x_range_high - x_range_low))) + 1) * float(
-                    #                 pixel_ratio))
-                    #             scale_val = scale_val if scale_val > 0 else 1
-                    #             scale_annot = str(scale_val) + "μm"
-                    #             scale_text = f'<span style="color: white">{str(scale_annot)}</span><br>'
-                    #             # get the index of the list element corresponding to this text annotation
-                    #             index = cur_graph['layout']['annotations'].index(annotations)
-                    #             cur_graph['layout']['annotations'][index]['text'] = scale_text
-                    #
-                    #             fig = go.Figure(cur_graph)
-                    #             fig.update_layout(newshape=dict(line=dict(color="white")))
-                    fig = CanvasLayout(cur_graph).update_scalebar_zoom_value(cur_graph_layout, pixel_ratio)
-                    return fig, cur_graph_layout
+                    # TODO: add ability to include the proportion here for custom scalebar lengths
+                    image_shape = get_first_image_from_roi_dictionary(image_dict[data_selection]).shape
+                    proportion = float(custom_scale_val / image_shape[1]) if custom_scale_val is not None else 0.075
+                    # if ctx.triggered_id == 'annotation_canvas':
+                    cur_graph = CanvasLayout(cur_graph).update_scalebar_zoom_value(cur_graph_layout, pixel_ratio, proportion)
+                    if ctx.triggered_id == "custom-scale-val":
+                        pixel_ratio = pixel_ratio if pixel_ratio is not None else 1
+                        x_axis_placement = set_x_axis_placement_of_scalebar(image_shape[1], invert_annot)
+                        cur_graph = CanvasLayout(cur_graph).toggle_scalebar(toggle_scalebar, x_axis_placement, invert_annot,
+                                                        pixel_ratio, image_shape, legend_size, proportion)
+                    # elif ctx.triggered_id == 'pixel-size-ratio':
+                    #     cur_graph = CanvasLayout(cur_graph).use_custom_scalebar_value(custom_scale_val, pixel_ratio, proportion)
+                    return cur_graph, cur_graph_layout
                 except (ValueError, KeyError, AssertionError):
                     raise PreventUpdate
             if ctx.triggered_id == "activate-coord":
                 if None not in (x_request, y_request, current_window) and \
                         nclicks_coord is not None and nclicks_coord > 0:
                     try:
-                        # # calculate midway distance for each coord. this distance is
-                        # # added on either side of the x and y requests
-                        # new_x_low, new_x_high, new_y_low, new_y_high = create_new_coord_bounds(current_window,
-                        #                                                                        x_request,
-                        #                                                                        y_request)
-                        # new_layout = {'xaxis.range[0]': new_x_low, 'xaxis.range[1]': new_x_high,
-                        #               'yaxis.range[0]': new_y_high, 'yaxis.range[1]': new_y_low}
-                        # # IMP: for yaxis, need to set the min and max in the reverse order
-                        # fig = go.Figure(data=cur_graph['data'], layout=cur_graph['layout'])
-                        # shapes = cur_graph['layout']['shapes']
-                        # annotations = cur_graph['layout']['annotations']
-                        # fig['layout']['shapes'] = None
-                        # fig['layout']['annotations'] = None
-                        # fig.update_layout(xaxis=XAxis(showticklabels=False, range=[new_x_low, new_x_high]),
-                        #                   yaxis=YAxis(showticklabels=False, range=[new_y_high, new_y_low]))
-                        # fig.update_layout(newshape=dict(line=dict(color="white")))
-                        # # cur_graph['layout']['xaxis']['domain'] = [0, 1]
-                        # # cur_graph['layout']['dragmode'] = "zoom"
-                        # fig['layout']['shapes'] = shapes
-                        # fig['layout']['annotations'] = annotations
-                        # return fig, new_layout
                         fig, new_layout = CanvasLayout(cur_graph).update_coordinate_window(current_window,
                                                                                      x_request, y_request)
                         return fig, new_layout
@@ -1200,46 +1149,25 @@ def init_pixel_level_callbacks(dash_app, tmpdirname, authentic_id, app_config):
     #     else:
     #         raise PreventUpdate
 
-    @dash_app.callback(Output('annotation_canvas', 'figure', allow_duplicate=True),
-                       State('annotation_canvas', 'figure'),
-                       State('annotation_canvas', 'relayoutData'),
-                       Input('custom-scale-val', 'value'),
-                       Input('pixel-size-ratio', 'value'),
-                       prevent_initial_call=True)
-    # @cache.memoize())
-    def render_canvas_from_scalebar_change(cur_graph, cur_graph_layout, custom_scale_val, pixel_ratio):
-        # do not update the canvas if the pixel ratio is changed to None
-        pixel_ratio_none = ctx.triggered_id == 'pixel-size-ratio' and pixel_ratio is None
-        if cur_graph is not None and cur_graph_layout not in [{'dragmode': 'pan'}] and not pixel_ratio_none:
-            try:
-                # cur_graph = strip_invalid_shapes_from_graph_layout(cur_graph)
-                # pixel_ratio = pixel_ratio if pixel_ratio is not None else 1
-                # for annotations in cur_graph['layout']['annotations']:
-                #     # if 'μm' in annotations['text'] and annotations['y'] == 0.06:
-                #     if annotations['y'] == 0.06:
-                #         if custom_scale_val is None:
-                #             high = max(cur_graph['layout']['xaxis']['range'][1],
-                #                        cur_graph['layout']['xaxis']['range'][0])
-                #             low = min(cur_graph['layout']['xaxis']['range'][1],
-                #                       cur_graph['layout']['xaxis']['range'][0])
-                #             x_range_high = math.ceil(int(high))
-                #             x_range_low = math.floor(int(low))
-                #             assert x_range_high >= x_range_low
-                #             custom_scale_val = int(float(math.ceil(int(0.075 *
-                #                                 (x_range_high - x_range_low))) + 1) * float(pixel_ratio))
-                #         scale_annot = str(custom_scale_val) + "μm"
-                #         scale_text = f'<span style="color: white">{str(scale_annot)}</span><br>'
-                #         # get the index of the list element corresponding to this text annotation
-                #         index = cur_graph['layout']['annotations'].index(annotations)
-                #         cur_graph['layout']['annotations'][index]['text'] = scale_text
-                # fig = go.Figure(cur_graph)
-                # fig.update_layout(newshape=dict(line=dict(color="white")))
-                fig = CanvasLayout(cur_graph).use_custom_scalebar_value(custom_scale_val, pixel_ratio)
-            except (KeyError, AssertionError):
-                fig = dash.no_update
-            return fig
-        else:
-            raise PreventUpdate
+    # @dash_app.callback(Output('annotation_canvas', 'figure', allow_duplicate=True),
+    #                    State('annotation_canvas', 'figure'),
+    #                    State('annotation_canvas', 'relayoutData'),
+    #                    Input('custom-scale-val', 'value'),
+    #                    Input('pixel-size-ratio', 'value'),
+    #                    prevent_initial_call=True)
+    # # @cache.memoize())
+    # def render_canvas_from_scalebar_change(cur_graph, cur_graph_layout, custom_scale_val, pixel_ratio):
+    #     # do not update the canvas if the pixel ratio is changed to None
+    #     pixel_ratio_none = ctx.triggered_id == 'pixel-size-ratio' and pixel_ratio is None
+    #     if cur_graph is not None and cur_graph_layout not in [{'dragmode': 'pan'}] and not pixel_ratio_none:
+    #         try:
+    #             # TODO: change the scalebar length in conjunction with the value
+    #             fig = CanvasLayout(cur_graph).use_custom_scalebar_value(custom_scale_val, pixel_ratio)
+    #         except (KeyError, AssertionError):
+    #             fig = dash.no_update
+    #         return fig
+    #     else:
+    #         raise PreventUpdate
 
     @dash_app.callback(Output('annotation_canvas', 'figure', allow_duplicate=True),
                        Input('toggle-canvas-legend', 'value'),
@@ -1256,27 +1184,24 @@ def init_pixel_level_callbacks(dash_app, tmpdirname, authentic_id, app_config):
                        State('pixel-size-ratio', 'value'),
                        State('invert-annotations', 'value'),
                        Input('legend_orientation', 'value'),
+                       State('custom-scale-val', 'value'),
                        prevent_initial_call=True)
     def render_canvas_from_toggle_show_annotations(toggle_legend, toggle_scalebar,
                                                    cur_canvas, cur_layout, currently_selected,
                                                    data_selection, blend_colour_dict, aliases, image_dict,
                                                    channel_order, legend_size, pixel_ratio, invert_annot,
-                                                   legend_orientation):
+                                                   legend_orientation, custom_scale_val):
         """
-        re-render the canvas if the user requests to remove the annotations (scalebar and legend)
+        re-render the canvas if the user requests to remove the annotations (scalebar and legend) or
+        updates the scalebar length with a custom value
         """
         if None not in (cur_layout, cur_canvas, data_selection, currently_selected, blend_colour_dict):
             # scalebar is y = 0.06
             # legend is y = 0.05
             cur_canvas = strip_invalid_shapes_from_graph_layout(cur_canvas)
             pixel_ratio = pixel_ratio if pixel_ratio is not None else 1
-            first_image = get_first_image_from_roi_dictionary(image_dict[data_selection])
-            image_shape = first_image.shape
-            x_axis_placement = 0.00001 * image_shape[1]
-            # make sure the placement is min 0.05 and max 0.1
-            x_axis_placement = x_axis_placement if 0.05 <= x_axis_placement <= 0.15 else 0.05
-            if invert_annot:
-                x_axis_placement = 1 - x_axis_placement
+            image_shape = get_first_image_from_roi_dictionary(image_dict[data_selection]).shape
+            x_axis_placement = set_x_axis_placement_of_scalebar(image_shape[1], invert_annot)
             # if 'layout' in cur_canvas and 'annotations' in cur_canvas['layout']:
             #     cur_annotations = cur_canvas['layout']['annotations'].copy()
             # else:
@@ -1289,64 +1214,10 @@ def init_pixel_level_callbacks(dash_app, tmpdirname, authentic_id, app_config):
                 legend_text = generate_canvas_legend_text(blend_colour_dict,
                             channel_order, aliases, legend_orientation)  if toggle_legend else ''
                 return CanvasLayout(cur_canvas).toggle_legend(toggle_legend, legend_text, x_axis_placement, legend_size)
-                # cur_annotations = [annot for annot in cur_annotations if \
-                #                    annot is not None and 'y' in annot and annot['y'] != 0.05]
-                # cur_canvas['layout']['annotations'] = cur_annotations
-                # if not toggle_legend:
-                #     return cur_canvas
-                # else:
-                #     fig = go.Figure(cur_canvas)
-                #     fig.update_layout(newshape=dict(line=dict(color="white")))
-                #     if legend_text != '':
-                #         fig.add_annotation(text=legend_text, font={"size": legend_size + 1}, xref='paper',
-                #                                yref='paper',
-                #                                x=(1 - x_axis_placement),
-                #                                # xanchor='right',
-                #                                y=0.05,
-                #                                # yanchor='bottom',
-                #                                bgcolor="black",
-                #                                showarrow=False)
-                #     return fig
-            elif ctx.triggered_id == "toggle-canvas-scalebar":
-                # if not toggle_scalebar:
-                #     cur_shapes = [shape for shape in cur_shapes if \
-                #                       shape is not None and 'type' in shape and shape['type'] \
-                #                       in ['rect', 'path', 'circle']]
-                #     cur_annotations = [annot for annot in cur_annotations if \
-                #                            annot is not None and 'y' in annot and annot['y'] != 0.06]
-                #     cur_canvas['layout']['annotations'] = cur_annotations
-                #     cur_canvas['layout']['shapes'] = cur_shapes
-                #     return cur_canvas
-                # else:
-                #     fig = go.Figure(cur_canvas)
-                #     # set the x0 and x1 depending on if the bar is inverted or not
-                #     x_0 = x_axis_placement if not invert_annot else (x_axis_placement - 0.075)
-                #     x_1 = (x_axis_placement + 0.075) if not invert_annot else x_axis_placement
-                #     fig.add_shape(type="line",
-                #                   xref="paper", yref="paper",
-                #                   x0=x_0, y0=0.05, x1=x_1,
-                #                   y1=0.05, line=dict(color="white", width=2))
-                #
-                #     try:
-                #         high = max(cur_canvas['layout']['xaxis']['range'][1],
-                #                    cur_canvas['layout']['xaxis']['range'][0])
-                #         low = min(cur_canvas['layout']['xaxis']['range'][1],
-                #                   cur_canvas['layout']['xaxis']['range'][0])
-                #         x_range_high = math.ceil(int(high))
-                #         x_range_low = math.floor(int(low))
-                #         assert x_range_high >= x_range_low
-                #         custom_scale_val = int(float(math.ceil(int(0.075 *
-                #                                                    (x_range_high - x_range_low))) + 1) * float(
-                #             pixel_ratio))
-                #     except KeyError:
-                #         custom_scale_val = None
-                #
-                #     fig = add_scale_value_to_figure(fig, image_shape, scale_value=custom_scale_val,
-                #                                     font_size=legend_size, x_axis_left=x_axis_placement,
-                #                                     invert=invert_annot)
-                # fig.update_layout(newshape=dict(line=dict(color="white")))
+            elif ctx.triggered_id in ["toggle-canvas-scalebar"]:
+                proportion = float(custom_scale_val / image_shape[1]) if custom_scale_val is not None else 0.075
                 return CanvasLayout(cur_canvas).toggle_scalebar(toggle_scalebar, x_axis_placement, invert_annot,
-                                                                pixel_ratio, image_shape, legend_size)
+                                                                pixel_ratio, image_shape, legend_size, proportion)
         else:
             raise PreventUpdate
 
@@ -1377,16 +1248,6 @@ def init_pixel_level_callbacks(dash_app, tmpdirname, authentic_id, app_config):
         """
         if cur_graph is not None:
             try:
-                # annotations_copy = cur_graph['layout']['annotations'].copy()
-                # for annotation in annotations_copy:
-                #     # the scalebar is always slightly smaller
-                #     if annotation['y'] == 0.06:
-                #         annotation['font']['size'] = legend_size
-                #     elif annotation['y'] == 0.05 and 'color' in annotation['text']:
-                #         annotation['font']['size'] = legend_size + 1
-                # cur_graph['layout']['annotations'] = [elem for elem in cur_graph['layout']['annotations'] if \
-                #                                   elem is not None and 'texttemplate' not in elem]
-                # return cur_graph
                 return CanvasLayout(cur_graph).change_annotation_size(legend_size)
             except KeyError:
                 raise PreventUpdate
@@ -1659,9 +1520,9 @@ def init_pixel_level_callbacks(dash_app, tmpdirname, authentic_id, app_config):
                           ["data-collection", "uploaded_dict", "annotation_canvas"] and \
                 active_tab == 'gallery-tab'
             use_zoom = gallery_data is not None and ctx.triggered_id == 'annotation_canvas'
-            if new_collection or gallery_mod_in_tab or use_zoom:
+            zoom_not_needed = ctx.triggered_id == 'annotation_canvas' and not toggle_gallery_zoom
+            if (new_collection or gallery_mod_in_tab or use_zoom) and not zoom_not_needed:
                 zoom_keys = ['xaxis.range[1]', 'xaxis.range[0]', 'yaxis.range[1]', 'yaxis.range[0]']
-                views = None
                 # maintain the original order of channels that is dictated by the metadata
                 # decide if channel view or ROI view is selected
                 # channel view
@@ -1725,7 +1586,7 @@ def init_pixel_level_callbacks(dash_app, tmpdirname, authentic_id, app_config):
                             label = aliases[key] if aliases is not None and key in aliases.keys() else key
                             cell_styling_conditions.append( {"condition": f"params.value == '{label}'",
                             "style": {"color": f"{blend_colours[key]['color']}"}})
-                    except KeyError as e:
+                    except KeyError:
                         pass
                 if len(in_blend) > 0:
                     to_return = pd.DataFrame(in_blend, columns=["Channel"]).to_dict(orient="records")
@@ -1936,7 +1797,6 @@ def init_pixel_level_callbacks(dash_app, tmpdirname, authentic_id, app_config):
         Reset the range slider for the current channel to the default values (min of 0 and max of 99th pixel
         percentile)
         """
-        # TODO: figure out why the values do not update properly when the max is less than 1
         if None not in (selected_channel, uploaded, data_selection, current_blend_dict) and reset > 0:
             hist_max = upper_bound_for_range_slider(uploaded[data_selection][selected_channel])
             upper_bound = float(get_default_channel_upper_bound_by_percentile(uploaded[data_selection][selected_channel]))